add image folder + divide javascript functions
