+++ mong.html
@@ -12,6 +12,6 @@
             <div class = "part2"><p>Log In&nbsp;&nbsp;&nbsp;</p></div>
         </div>
         <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-        <script src="functions.js"></script>
+        <script type = "module" src="functions.js"></script>
     </body>
 </html>
